+++ R/timeAverage.R
@@ -357,7 +357,8 @@ timeAverage <- function(mydata, avg.time = "day", data.thresh = 0,
       allData <- data.frame(date = seq(min(allDates), max(allDates), avg.time))
       
       ## merge with orginal data, which leaves gaps to fill
-      mydata <- full_join(mydata, allData, by = "date")
+      mydata <- full_join(mydata, allData, by = "date") %>%
+        arrange(date)
       
       if (fill) {