+++ NEWS.md
@@ -1,6 +1,7 @@
 # Changes in openair 1.9-99
 
 - Only plot lower triangle for `corPlot`
+- make sure date class is POSIXct, POSIXt in `importAURN`
 
 # Changes in openair 1.8-2