+++ NEWS.md
@@ -6,6 +6,7 @@
 - fix `pollutionRose` plot warning
 - fix NMB in `modStats`
 - fix bug in `timeAverage` where estimating the time interval in input data could be unreliable due to low data availability
+- fix time expansion bug in `timeAverage`
 
 # Changes in openair 1.8-2